# Entities/DTO/Output/BuildingOutput.cs
namespace Entities.DTO.Output
{
    public class BuildingOutput
    {
        public Guid Id { get; set; }
        public string Name { get; set; }
        public string Address { get; set; }
        public DateTime Updated_At { get; set; }

    }
}